Added rings + owls with a flock/follow target behaviour
- A list of rings spawn, if the player goes through one, it gets
destroyed and a new one instantiates somewhere random. The "points"
feedback is a sound effect as the OnGUI() does not work with the Oculus
Rift cameras. Regardless, there is a point counter (+10 per ring)

- In the world that David created, the lights were missing, I added them

- Somehow the moth wings .ma file did not import into Unity (I am
actually not sure whether Unity reads that file extension); so I went
into Maya and exported the wings as .fbx and reimported them into the
scene.

- I added owls that act as a flock. They follow a target that is
constantly moving

- I disabled the Bird's cube mesh renderer as we don't want to see a
cube instead of some creature. I also changed the collider to a sphere
so that it is smoother when colliding with one of the rings.

- The wings look great but they should appear in front of the camera
(if not, no one is going to see them when in game mode). They are still
in (0,0,0) as it doesn't have a body and it would look slightly weird...